+++ pcs/lib/commands/resource.py
@@ -22,6 +22,7 @@ from pcs.lib.errors import LibraryError
 from pcs.lib.pacemaker.values import validate_id
 from pcs.lib.pacemaker.state import (
     ensure_resource_state,
+    info_resource_state,
     is_resource_managed,
     ResourceNotFound,
 )
@@ -31,7 +32,10 @@ from pcs.lib.resource_agent import(
 
 @contextmanager
 def resource_environment(
-    env, wait=False, wait_for_resource_ids=None, disabled_after_wait=False,
+    env,
+    wait=False,
+    wait_for_resource_ids=None,
+    resource_state_reporter=info_resource_state,
     required_cib_version=None
 ):
     env.ensure_wait_satisfiable(wait)
@@ -41,9 +45,18 @@ def resource_environment(
     if wait is not False and wait_for_resource_ids:
         state = env.get_cluster_state()
         env.report_processor.process_list([
-            ensure_resource_state(not disabled_after_wait, state, res_id)
+            resource_state_reporter(state, res_id)
             for res_id in wait_for_resource_ids
         ])
+
+def _ensure_disabled_after_wait(disabled_after_wait):
+    def inner(state, resource_id):
+        return ensure_resource_state(
+            not disabled_after_wait,
+            state,
+            resource_id
+        )
+    return inner
 
 def _validate_remote_connection(
     resource_agent, nodes_to_validate_against, resource_id, instance_attributes,
@@ -195,7 +208,11 @@ def create(
         env,
         wait,
         [resource_id],
-        ensure_disabled or resource.common.are_meta_disabled(meta_attributes),
+        _ensure_disabled_after_wait(
+            ensure_disabled
+            or
+            resource.common.are_meta_disabled(meta_attributes)
+        )
     ) as resources_section:
         _check_special_cases(
             env,
@@ -269,7 +286,7 @@ def _create_as_clone_common(
         env,
         wait,
         [resource_id],
-        (
+        _ensure_disabled_after_wait(
             ensure_disabled
             or
             resource.common.are_meta_disabled(meta_attributes)
@@ -353,7 +370,11 @@ def create_in_group(
         env,
         wait,
         [resource_id],
-        ensure_disabled or resource.common.are_meta_disabled(meta_attributes),
+        _ensure_disabled_after_wait(
+            ensure_disabled
+            or
+            resource.common.are_meta_disabled(meta_attributes)
+        )
     ) as resources_section:
         _check_special_cases(
             env,
@@ -433,7 +454,11 @@ def create_into_bundle(
         env,
         wait,
         [resource_id],
-        disabled_after_wait=ensure_disabled,
+        _ensure_disabled_after_wait(
+            ensure_disabled
+            or
+            resource.common.are_meta_disabled(meta_attributes)
+        ),
         required_cib_version=(2, 8, 0)
     ) as resources_section:
         _check_special_cases(
@@ -465,8 +490,9 @@ def create_into_bundle(
 
 def bundle_create(
     env, bundle_id, container_type, container_options=None,
-    network_options=None, port_map=None, storage_map=None,
+    network_options=None, port_map=None, storage_map=None, meta_attributes=None,
     force_options=False,
+    ensure_disabled=False,
     wait=False,
 ):
     """
@@ -477,24 +503,32 @@ def bundle_create(
     string container_type -- container engine name (docker, lxc...)
     dict container_options -- container options
     dict network_options -- network options
-    list of dict port_map -- list of port mapping options
-    list of dict storage_map -- list of storage mapping options
+    list of dict port_map -- a list of port mapping options
+    list of dict storage_map -- a list of storage mapping options
+    dict meta_attributes -- bundle's meta attributes
     bool force_options -- return warnings instead of forceable errors
+    bool ensure_disabled -- set the bundle's target-role to "Stopped"
     mixed wait -- False: no wait, None: wait default timeout, int: wait timeout
     """
     container_options = container_options or {}
     network_options = network_options or {}
     port_map = port_map or []
     storage_map = storage_map or []
+    meta_attributes = meta_attributes or {}
 
     with resource_environment(
         env,
         wait,
         [bundle_id],
-        # bundles are always enabled, currently there is no way to disable them
-        disabled_after_wait=False,
+        _ensure_disabled_after_wait(
+            ensure_disabled
+            or
+            resource.common.are_meta_disabled(meta_attributes)
+        ),
         required_cib_version=(2, 8, 0)
     ) as resources_section:
+        # no need to run validations related to remote and guest nodes as those
+        # nodes can only be created from primitive resources
         id_provider = IdProvider(resources_section)
         env.report_processor.process_list(
             resource.bundle.validate_new(
@@ -505,10 +539,11 @@ def bundle_create(
                 network_options,
                 port_map,
                 storage_map,
+                # TODO meta attributes - there is no validation for now
                 force_options
             )
         )
-        resource.bundle.append_new(
+        bundle_element = resource.bundle.append_new(
             resources_section,
             id_provider,
             bundle_id,
@@ -516,13 +551,16 @@ def bundle_create(
             container_options,
             network_options,
             port_map,
-            storage_map
+            storage_map,
+            meta_attributes
         )
+        if ensure_disabled:
+            resource.common.disable(bundle_element)
 
 def bundle_update(
     env, bundle_id, container_options=None, network_options=None,
     port_map_add=None, port_map_remove=None, storage_map_add=None,
-    storage_map_remove=None,
+    storage_map_remove=None, meta_attributes=None,
     force_options=False,
     wait=False,
 ):
@@ -537,6 +575,7 @@ def bundle_update(
     list of string port_map_remove -- list of port mapping ids to remove
     list of dict storage_map_add -- list of storage mapping options to add
     list of string storage_map_remove -- list of storage mapping ids to remove
+    dict meta_attributes -- meta attributes to update
     bool force_options -- return warnings instead of forceable errors
     mixed wait -- False: no wait, None: wait default timeout, int: wait timeout
     """
@@ -546,15 +585,16 @@ def bundle_update(
     port_map_remove = port_map_remove or []
     storage_map_add = storage_map_add or []
     storage_map_remove = storage_map_remove or []
+    meta_attributes = meta_attributes or {}
 
     with resource_environment(
         env,
         wait,
         [bundle_id],
-        # bundles are always enabled, currently there is no way to disable them
-        disabled_after_wait=False,
         required_cib_version=(2, 8, 0)
     ) as resources_section:
+        # no need to run validations related to remote and guest nodes as those
+        # nodes can only be created from primitive resources
         id_provider = IdProvider(resources_section)
         bundle_element = find_element_by_tag_and_id(
             resource.bundle.TAG,
@@ -571,6 +611,7 @@ def bundle_update(
                 port_map_remove,
                 storage_map_add,
                 storage_map_remove,
+                # TODO meta attributes - there is no validation for now
                 force_options
             )
         )
@@ -582,7 +623,8 @@ def bundle_update(
             port_map_add,
             port_map_remove,
             storage_map_add,
-            storage_map_remove
+            storage_map_remove,
+            meta_attributes
         )
 
 def disable(env, resource_ids, wait):
@@ -593,7 +635,7 @@ def disable(env, resource_ids, wait):
     mixed wait -- False: no wait, None: wait default timeout, int: wait timeout
     """
     with resource_environment(
-        env, wait, resource_ids, True
+        env, wait, resource_ids, _ensure_disabled_after_wait(True)
     ) as resources_section:
         resource_el_list = _find_resources_or_raise(
             resources_section,
@@ -615,7 +657,7 @@ def enable(env, resource_ids, wait):
     mixed wait -- False: no wait, None: wait default timeout, int: wait timeout
     """
     with resource_environment(
-        env, wait, resource_ids, False
+        env, wait, resource_ids, _ensure_disabled_after_wait(False)
     ) as resources_section:
         resource_el_list = _find_resources_or_raise(
             resources_section,
@@ -642,7 +684,7 @@ def _resource_list_enable_disable(resource_el_list, func, cluster_state):
             report_list.append(
                 reports.id_not_found(
                     res_id,
-                    id_description="resource/clone/master/group"
+                    id_description="resource/clone/master/group/bundle"
                )
             )
     return report_list
@@ -735,7 +777,7 @@ def _find_resources_or_raise(
     resource_tags = (
         resource.clone.ALL_TAGS
         +
-        [resource.group.TAG, resource.primitive.TAG]
+        [resource.group.TAG, resource.primitive.TAG, resource.bundle.TAG]
     )
     for res_id in resource_ids:
         try:
@@ -745,7 +787,7 @@ def _find_resources_or_raise(
                         resource_tags,
                         resources_section,
                         res_id,
-                        id_description="resource/clone/master/group"
+                        id_description="resource/clone/master/group/bundle"
                     )
                 )
             )