+++ pcs/usage.py
@@ -430,10 +430,12 @@ Commands:
 
     bundle create <bundle id> [container [<container type>] <container options>]
             [network <network options>] [port-map <port options>]...
-            [storage-map <storage options>]... [--wait[=n]]
+            [storage-map <storage options>]... [meta <meta options>]
+            [--disabled] [--wait[=n]]
         Create a new bundle encapsulating no resources. The bundle can be used
         either as it is or a resource may be put into it at any time.
         If the container type is not specified, it defaults to 'docker'.
+        If --disabled is specified, the bundle is not started automatically.
         If --wait is specified, pcs will wait up to 'n' seconds for the bundle
         to start and then return 0 on success or 1 on error. If 'n' is not
         specified it defaults to 60 minutes.
@@ -442,13 +444,14 @@ Commands:
             [network <network options>]
             [port-map (add <port options>) | (remove <id>...)]...
             [storage-map (add <storage options>) | (remove <id>...)]...
+            [meta <meta options>]
             [--wait[=n]]
         Add, remove or change options to specified bundle. If you wish to update
         a resource encapsulated in the bundle, use the 'pcs resource update'
-        command instead and specify the resource id.  If --wait is specified,
+        command instead and specify the resource id. If --wait is specified,
         pcs will wait up to 'n' seconds for the operation to finish (including
         moving resources if appropriate) and then return 0 on success or 1 on
-        error.  If 'n' is not specified it defaults to 60 minutes.
+        error. If 'n' is not specified it defaults to 60 minutes.
 
     manage <resource id>... [--monitor]
         Set resources listed to managed mode (default). If --monitor is